+++ android/app/build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
     id 'kotlinx-serialization'
-    id 'kotlin-kapt'
+//    id 'kotlin-kapt'
 }
 
 android {
