+++ android/app/src/main/java/com/digginroom/digginroom/viewmodels/JoinViewModel.kt
@@ -1,7 +1,5 @@
-package com.digginroom.digginroom
+package com.digginroom.digginroom.viewmodels
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
@@ -12,33 +10,36 @@ import com.digginroom.model.user.User
 import com.digginroom.repository.UserRepository
 
 class JoinViewModel(
-    private val userRepository: UserRepository
-//    private val joinSucceed: () -> Unit,
-//    private val joinFailed: () -> Unit
+    private val userRepository: UserRepository,
+    private val joinSucceed: () -> Unit,
+    private val joinFailed: () -> Unit
 ) : ViewModel() {
 
-    private val _state: MutableLiveData<JoinState> = MutableLiveData()
-    val state: LiveData<JoinState>
-        get() = _state
-
     fun join(id: String, password: String) {
-        _state.value = userRepository.save(
+        val joinState = userRepository.save(
             User(
                 id = id,
                 password = password
             )
         )
+        when (joinState) {
+            is JoinState.Success -> joinSucceed()
+            is JoinState.Retry -> joinFailed()
+        }
     }
 
     companion object {
 
-        fun getJoinViewModelFactory(): ViewModelProvider.Factory {
+        fun getJoinViewModelFactory(
+            joinSucceed: () -> Unit,
+            joinFailed: () -> Unit
+        ): ViewModelProvider.Factory {
             return viewModelFactory {
                 initializer {
                     JoinViewModel(
-                        userRepository = DefaultUserRepository()
-//                        joinSucceed = joinSucceed,
-//                        joinFailed = joinFailed
+                        userRepository = DefaultUserRepository(),
+                        joinSucceed = joinSucceed,
+                        joinFailed = joinFailed
                     )
                 }
             }
